refactor: extract shared DtcPageBase and fix code review issues
- Extract shared DtcPageBase component; ClientDtcPage and DtcPage are
  now thin wrappers (~15 lines each) eliminating ~90 lines of duplication
- scenario-5/6: import apiUrl() from actuaries/helpers instead of
  redefining it locally in each file
- RecurringOrderModal: add aria-labelledby pointing to modal h3 title
