+++ cypress/e2e/actuaries/scenario-5-employee-birth-date-future.cy.ts
@@ -1,12 +1,7 @@
 /// <reference types="cypress" />
 
 import { fillInputByLabel, fillDateByLabel, selectByLabel } from '../../support/formByLable';
-
-function apiUrl() {
-    const url = Cypress.env('API_URL') as string | undefined;
-    const fallbackUrl = 'http://rafsi.davidovic.io:8080/api';
-    return url && !url.includes('localhost') ? url : fallbackUrl;
-}
+import { apiUrl } from './helpers';
 
 function fillValidEmployeeForm(dateOfBirth: string) {
     const ts = Date.now();
